pyproject.toml
[project]
name = "e2e-taxi-ride-duration-prediction"
version = "0.1.0"
description = "An end-to-end MLOps project showcasing a complete machine learning pipeline using the NYC Taxi dataset. The goal is to predict taxi ride durations based on trip data and demonstrate modern MLOps practices, from data ingestion to model deployment and monitoring."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "evidently>=0.7.11",
  "fastapi[standard]>=0.115.14",
  "ipykernel>=6.29.5",
  "ipywidgets>=8.1.7",
  "joblib>=1.5.1",
  "loguru>=0.7.3",
  "mlflow>=3.1.4",
  "polars>=1.31.0",
  "prefect>=3.4.11",
  "pyarrow<=19.0.1",
  "pydantic>=2.11.7",
  "requests>=2.32.4",
  "scikit-learn>=1.7.1",
  "seaborn>=0.13.2",
  "tqdm>=4.67.1",
  "xgboost>=3.0.2",
]

[project.optional-dependencies]
nvidia = ["cuml-cu12==25.6.*", "pandas>=2.0,<2.2.4"]
profiling = []

[dependency-groups]
dev = [
  "gitlint>=0.19.1",
  "jupyter>=1.1.1",
  "nbmake>=1.5.5",
  "pre-commit>=4.2.0",
  "pytest>=8.4.1",
  "pytest-cov>=6.2.1",
  "ruff>=0.12.1",
  "scipy-stubs>=1.16.0.2",
]

[build-system]
requires = ["uv_build>=0.7.19, <0.8.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = ""
module-name = "e2e_taxi_ride_duration_prediction"

[[tool.uv.index]]
url = "https://pypi.nvidia.com/"
